refactor of downloader and parsing and agg

+++ capstone/etl/eia_retrieval.py
@@ -1,13 +1,16 @@
 import urllib.request
+from pathlib import Path
 
 
 def eia_retrieval(workspace_directory):
-    url = "https://www.eia.gov/petroleum/drilling/xls/"
     filename = "dpr-data.xlsx"
 
-    urllib.request.urlretrieve(
-        f"{url}/{filename}",
-        f"{workspace_directory}/{filename}",
-    )
+    if not Path(f"{workspace_directory}/{filename}").is_file():
+        url = "https://www.eia.gov/petroleum/drilling/xls/"
+
+        urllib.request.urlretrieve(
+            f"{url}/{filename}",
+            f"{workspace_directory}/{filename}",
+        )
 
     return f"{workspace_directory}/{filename}"
